Projeto em formato maven e add testes
