+++ Source Code Inspection/src/main/java/com/mycompany/ticketmaster/TicketMaster.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ */
+
+package com.mycompany.ticketmaster;
+
+/**
+ *
+ * @author jehsi
+ */
+public class TicketMaster {
+
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
