+++ Source Code Inspection/src/main/java/com/mycompany/ticketmaster/core/PapelMoeda.java
@@ -1,0 +1,24 @@
+package com.mycompany.ticketmaster.core;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoeda {
+
+    protected int valor;
+    protected int quantidade;
+
+    public PapelMoeda(int valor, int quantidade) {
+        this.valor = valor;
+        this.quantidade = quantidade;
+    }
+
+    public int getValor() {
+        return valor;
+    }
+
+    public int getQuantidade() {
+        return quantidade;
+    }
+}
