+++ Source Code Inspection/src/main/java/com/mycompany/ticketmaster/core/Troco.java
@@ -1,0 +1,64 @@
+package com.mycompany.ticketmaster.core;
+
+import java.util.Iterator;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+class Troco {
+
+    protected PapelMoeda[] papeisMoeda;
+
+    public Troco(int valor) {
+        papeisMoeda = new PapelMoeda[6];
+        int[] valoresMoedas = {100, 50, 20, 10, 5, 2};
+
+        for (int i = 0; i < valoresMoedas.length; i++) {
+            int count = valor / valoresMoedas[i];
+            valor %= valoresMoedas[i];
+            papeisMoeda[i] = new PapelMoeda(valoresMoedas[i], count);
+   
+        }
+    }
+
+    public Iterator<PapelMoeda> getIterator() {
+        return new TrocoIterator(this);
+    }
+
+    class TrocoIterator implements Iterator<PapelMoeda> {
+
+        protected Troco troco;
+
+        public TrocoIterator(Troco troco) {
+            this.troco = troco;
+        }
+
+        @Override
+        public boolean hasNext() {
+            for (int i = 0; i <= 5; i++) {
+                if (troco.papeisMoeda[i] != null) {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        @Override
+        public PapelMoeda next() {
+            PapelMoeda ret = null;
+            for (int i = 0; i <= 5 && ret != null; i++) {
+                if (troco.papeisMoeda[i] != null) {
+                    ret = troco.papeisMoeda[i];
+                    troco.papeisMoeda[i] = null;
+                }
+            }
+            return ret;
+        }
+
+        @Override
+        public void remove() {
+            next();
+        }
+    }
+}
