+++ Source Code Inspection/src/main/java/com/mycompany/ticketmaster/exception/PapelMoedaInvalidaException.java
@@ -1,0 +1,9 @@
+package com.mycompany.ticketmaster.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class PapelMoedaInvalidaException extends Exception {
+    
+}
