+++ Source Code Inspection/src/main/java/com/mycompany/ticketmaster/exception/SaldoInsuficienteException.java
@@ -1,0 +1,8 @@
+package com.mycompany.ticketmaster.exception;
+
+/**
+ *
+ * @author Calebe de Paula Bianchini
+ */
+public class SaldoInsuficienteException extends Exception {
+}
